타입 오류 수정
- _selectedSound!.length를 int로 캐스팅하여 타입 오류 해결
- 자동 단가 계산 기능이 정상적으로 작동하도록 수정

+++ lib/features/post_system/screens/post_place_screen.dart
@@ -184,7 +184,7 @@ class _PostPlaceScreenState extends State<PostPlaceScreen> {
     
     // 사운드 크기 계산
     if (_selectedSound != null) {
-      totalSize += _selectedSound!.length;
+      totalSize += _selectedSound!.length.toInt();
     }
     
     // 단가 계산
